split feature generation from model training module
